+++ Sources/BottomSheet/Examples/RainbowList.swift
@@ -1,0 +1,27 @@
+//
+//  RainbowList.swift
+//  BottomSheet
+//
+//  Created by Sam Doggett on 10/28/24.
+//
+
+import SwiftUI
+
+struct RainbowList: View {
+
+    let colors: [Color] = [.red, .orange, .yellow, .green, .blue, .indigo, .purple]
+
+    var body: some View {
+        List {
+            ForEach((0..<colors.count * 4), id: \.self) { index in
+                let color = colors[index % colors.count]
+                Text(color.description)
+                    .foregroundColor(.white)
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .listRowBackground(color)
+            }
+        }
+        .listStyle(.plain)
+    }
+
+}
